requirements.txt
python-dotenv
pyspark==3.5.1
json
time
asyncio
aiohttp
streamlit
aiokafka
